+++ Example/index.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import Main from './app/Main';
+
+AppRegistry.registerComponent('Example', () => Main);